feat: Move SynAppTestTags to core model module
This commit refactors the `SynAppTestTags` object, moving it from the `app` module to the `core:model` module. This allows the test tags to be shared and reused in other modules, particularly in the `app` module's UI tests.

The `SynAppTest.kt` is updated to import the `SynAppTestTags` from its new location and to use UI interactions (clicks) for navigation instead of directly manipulating the NavController state.
