+++ core/model/src/commonMain/kotlin/com/mshdabiola/model/testtag/SynAppTestTags.kt
@@ -1,0 +1,8 @@
+package com.mshdabiola.model.testtag
+
+object SynAppTestTags {
+    const val APP_ROOT_LAYOUT = "kmt_app:root_layout" // For SharedTransitionLayout or KmtBackground
+    const val GRADIENT_BACKGROUND = "kmt_app:gradient_background"
+    const val MAIN_SCAFFOLD = "kmt_app:main_scaffold" // Instance of KmtScaffold
+    const val NAV_HOST = "kmt_app:nav_host"
+}
